Add module for Digest sugar for the project
This module is meant to wrap the digest functions within the
project to allow the hashing to be changed in all places if
need, or minimally to control the interface for hashing.

Rather than import and define the same thing in various places,
or potentially import and define different hashing modules in
various places, it's defined here for continuity.