Fix logic to determine col type
Using the first row to determine all following is not accurate.
Additionally, the xlxs uses text abbreviations to denote meaning
within INTEGER and TEXT columns.  We also need to skip values that
are empty string unless the Image link column which we know are
always string but not present in the downloaded xlxs file.